+++ lib/ferrum.rb
@@ -1,5 +1,8 @@
 # frozen_string_literal: true
 
+require "concurrent-ruby"
+require "ferrum/utils/event"
+require "ferrum/utils/thread"
 require "ferrum/utils/platform"
 require "ferrum/utils/elapsed_time"
 require "ferrum/utils/attempt"